feat: Dev opens PR, removed Release stage, updated workflow graph
